backend/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# MinIO object storage
minio==7.2.3

# Database
supabase==2.3.0

# OpenAI and AI (OpenRouter compatible)
openai==1.10.0

# Vector database
chromadb==0.4.22

# Document processing
pypdf2==3.0.1
pdfplumber==0.10.3
pymupdf>=1.24.0
pillow>=10.2.0
pytesseract==0.3.10
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4

# Text processing
sentence-transformers==2.3.1
langdetect==1.0.9

# Utilities
requests==2.31.0
numpy==1.26.3
scikit-learn==1.4.0